Controller split
