+++ app/controllers/species_lists_controller.rb
@@ -201,96 +201,6 @@ class SpeciesListsController < ApplicationController
 
   ##############################################################################
   #
-  #  :section: Reports
-  #
-  ##############################################################################
-
-  def download
-    pass_query_params
-    @list = find_or_goto_index(SpeciesList, params[:id].to_s)
-    @type = params[:type] || "txt"
-    @format = params[:format] || "raw"
-    @encoding = params[:encoding] || "UTF-8"
-    @query = lookup_species_list_query(@list)
-  end
-
-  def print_labels
-    species_list = find_or_goto_index(SpeciesList, params[:id].to_s)
-    query = lookup_species_list_query(species_list)
-    redirect_with_query(print_labels_for_observations_path, query)
-  end
-
-  def lookup_species_list_query(list)
-    Query.lookup_and_save(:Observation, :in_species_list, species_list: list)
-  end
-
-  # Used by show_species_list.
-  def make_report
-    @species_list = find_or_goto_index(SpeciesList, params[:id].to_s)
-    return unless @species_list
-
-    names = @species_list.names
-    case params[:type]
-    when "txt"
-      render_name_list_as_txt(names)
-    when "rtf"
-      render_name_list_as_rtf(names)
-    when "csv"
-      render_name_list_as_csv(names)
-    else
-      flash_error(:make_report_not_supported.t(type: params[:type]))
-      redirect_to(action: "show_species_list", id: params[:id].to_s)
-    end
-  end
-
-  def render_name_list_as_txt(names)
-    charset = "UTF-8"
-    str = "\xEF\xBB\xBF#{names.map(&:real_search_name).join("\r\n")}"
-    send_data(str, type: "text/plain",
-                   charset: charset,
-                   disposition: "attachment",
-                   filename: "report.txt")
-  end
-
-  def render_name_list_as_csv(names)
-    charset = "ISO-8859-1"
-    str = CSV.generate do |csv|
-      csv << %w[scientific_name authority citation accepted]
-      names.each do |name|
-        csv << [name.real_text_name, name.author, name.citation,
-                name.deprecated ? "" : "1"].map(&:presence)
-      end
-    end
-    str = str.iconv(charset)
-    send_data(str, type: "text/csv",
-                   charset: charset,
-                   header: "present",
-                   disposition: "attachment",
-                   filename: "report.csv")
-  end
-
-  def render_name_list_as_rtf(names)
-    charset = "UTF-8"
-    doc = RTF::Document.new(RTF::Font::SWISS)
-    reportable_ranks = %w[Genus Species Subspecies Variety Form]
-    names.each do |name|
-      rank      = name.rank
-      text_name = name.real_text_name
-      author    = name.author
-      node      = name.deprecated ? doc : doc.bold
-      node      = node.italic if reportable_ranks.include?(rank)
-      node << text_name
-      doc << " #{author}" if author.present?
-      doc.line_break
-    end
-    send_data(doc.to_rtf, type: "text/rtf",
-                          charset: charset,
-                          disposition: "attachment",
-                          filename: "report.rtf")
-  end
-
-  ##############################################################################
-  #
   #  :section: Create and Modify
   #
   ##############################################################################
@@ -368,7 +278,7 @@ class SpeciesListsController < ApplicationController
   end
 
   # Form to let user create/edit species_list from file.
-  def upload_species_list
+  def upload
     @species_list = find_or_goto_index(SpeciesList, params[:id].to_s)
     return unless @species_list
 
@@ -413,292 +323,6 @@ class SpeciesListsController < ApplicationController
       flash_notice(:runtime_species_list_clear_success.t)
     end
     redirect_to(action: "show_species_list", id: @species_list)
-  end
-
-  def add_remove_observations
-    pass_query_params
-    @id = params[:species_list].to_s
-    @query = find_obs_query_or_redirect
-  end
-
-  def post_add_remove_observations
-    pass_query_params
-    id = params[:species_list].to_s
-    spl = find_list_or_reload_form(id)
-    return unless spl
-
-    query = find_obs_query_or_redirect(spl)
-    return unless query
-
-    do_add_remove_observations(spl, query)
-    redirect_to(action: "show_species_list", id: spl.id)
-  end
-
-  def find_obs_query_or_redirect(spl = nil)
-    query = find_query(:Observation)
-    return query if query
-
-    flash_error(:species_list_add_remove_no_query.t)
-    if spl
-      redirect_to(action: "show_species_list", id: spl.id)
-    else
-      redirect_to(action: "list_species_lists")
-    end
-    nil
-  end
-
-  def find_list_or_reload_form(id)
-    list = lookup_species_list_by_id_or_name(id)
-    return list if list
-
-    flash_error(:species_list_add_remove_bad_name.t(name: id.inspect))
-    redirect_to(add_query_param(action: :add_remove_observations,
-                                species_list: id))
-    nil
-  end
-
-  def lookup_species_list_by_id_or_name(str)
-    if /^\d+$/.match?(str)
-      SpeciesList.safe_find(str)
-    else
-      SpeciesList.find_by(title: str)
-    end
-  end
-
-  def do_add_remove_observations(spl, query)
-    return unless check_permission!(spl)
-
-    if params[:commit] == :ADD.l
-      do_add_observations(spl, query)
-    elsif params[:commit] == :REMOVE.l
-      do_remove_observations(spl, query)
-    else
-      flash_error("Invalid mode: #{params[:commit].inspect}")
-    end
-  end
-
-  def do_add_observations(species_list, query)
-    ids = query.result_ids - species_list.observation_ids
-    return if ids.empty?
-
-    # This is apparently extremely inefficient.  Danny says it times out for
-    # large species_lists, such as "Neotropical Fungi".
-    # species_list.observation_ids += ids
-    SpeciesListObservation.insert_all(
-      ids.map { |id| { observation_id: id, species_list_id: species_list.id } }
-    )
-    flash_notice(:species_list_add_remove_add_success.t(num: ids.length))
-  end
-
-  def do_remove_observations(species_list, query)
-    ids = query.result_ids & species_list.observation_ids
-    return if ids.empty?
-
-    species_list.observation_ids -= ids
-    flash_notice(:species_list_add_remove_remove_success.t(num: ids.length))
-  end
-
-  # Form to let user add/remove an observation from one of their species lists.
-  def manage_species_lists
-    @observation = find_or_goto_index(Observation, params[:id].to_s)
-    @all_lists = @user.all_editable_species_lists
-  end
-
-  # Used by manage_species_lists.
-  def remove_observation_from_species_list
-    species_list = find_or_goto_index(SpeciesList, params[:species_list])
-    return unless species_list
-
-    observation = find_or_goto_index(Observation, params[:observation])
-    return unless observation
-
-    if check_permission!(species_list)
-      species_list.remove_observation(observation)
-      flash_notice(:runtime_species_list_remove_observation_success.
-        t(name: species_list.unique_format_name, id: observation.id))
-      redirect_to(action: "manage_species_lists", id: observation.id)
-    else
-      redirect_to(action: "show_species_list", id: species_list.id)
-    end
-  end
-
-  # Used by manage_species_lists.
-  def add_observation_to_species_list
-    species_list = find_or_goto_index(SpeciesList, params[:species_list])
-    return unless species_list
-
-    observation = find_or_goto_index(Observation, params[:observation])
-    return unless observation
-
-    if check_permission!(species_list)
-      species_list.add_observation(observation)
-      flash_notice(:runtime_species_list_add_observation_success.
-        t(name: species_list.unique_format_name, id: observation.id))
-      redirect_to(action: "manage_species_lists", id: observation.id)
-    else
-      redirect_to(action: "show_species_list", id: species_list.id)
-    end
-  end
-
-  # ----------------------------
-  #  :section: Manage Projects
-  # ----------------------------
-
-  def manage_projects
-    return unless (@list = find_or_goto_index(SpeciesList, params[:id].to_s))
-
-    if check_permission!(@list)
-      @projects = projects_to_manage
-      @object_states = manage_object_states
-      @project_states = manage_project_states
-      if request.method == "POST"
-        if params[:commit] == :ATTACH.l
-          if attach_objects_to_projects
-            redirect_to(action: "show_species_list", id: @list.id)
-          else
-            flash_warning(:runtime_no_changes.t)
-          end
-        elsif params[:commit] == :REMOVE.l
-          if remove_objects_from_projects
-            redirect_to(action: "show_species_list", id: @list.id)
-          else
-            flash_warning(:runtime_no_changes.t)
-          end
-        else
-          flash_error("Invalid submit button: #{params[:commit].inspect}")
-        end
-      end
-    else
-      redirect_to(action: "show_species_list", id: @list.id)
-    end
-  end
-
-  def projects_to_manage
-    projects = @user.projects_member
-    if @list.user == @user
-      projects += @list.projects
-      projects.uniq!
-    end
-    projects
-  end
-
-  def manage_object_states
-    {
-      list: params[:objects_list].present?,
-      obs: params[:objects_obs].present?,
-      img: params[:objects_img].present?
-    }
-  end
-
-  def manage_project_states
-    states = {}
-    @projects.each do |proj|
-      states[proj.id] = params["projects_#{proj.id}"].present?
-    end
-    states
-  end
-
-  def attach_objects_to_projects
-    @any_changes = false
-    @projects.each do |proj|
-      if @project_states[proj.id]
-        if @user.projects_member.exclude?(proj)
-          flash_error(:species_list_projects_no_add_to_project.
-                         t(proj: proj.title))
-        else
-          attach_species_list_to_project(proj) if @object_states[:list]
-          attach_observations_to_project(proj) if @object_states[:obs]
-          attach_images_to_project(proj)       if @object_states[:img]
-        end
-      end
-    end
-    @any_changes
-  end
-
-  def remove_objects_from_projects
-    @any_changes = false
-    @projects.each do |proj|
-      next unless @project_states[proj.id]
-
-      remove_species_list_from_project(proj) if @object_states[:list]
-      remove_observations_from_project(proj) if @object_states[:obs]
-      remove_images_from_project(proj)       if @object_states[:img]
-    end
-    @any_changes
-  end
-
-  def attach_species_list_to_project(proj)
-    return if @list.projects.include?(proj)
-
-    proj.add_species_list(@list)
-    flash_notice(:attached_to_project.
-                    t(object: :species_list, project: proj.title))
-    @any_changes = true
-  end
-
-  def remove_species_list_from_project(proj)
-    return unless @list.projects.include?(proj)
-
-    proj.remove_species_list(@list)
-    flash_notice(:removed_from_project.
-                    t(object: :species_list, project: proj.title))
-    @any_changes = true
-  end
-
-  def attach_observations_to_project(proj)
-    obs = @list.observations.select { |o| check_permission(o) }
-    obs -= proj.observations
-    return unless obs.any?
-
-    proj.add_observations(obs)
-    flash_notice(:attached_to_project.
-                    t(object: "#{obs.length} #{:observations.l}",
-                      project: proj.title))
-    @any_changes = true
-  end
-
-  def remove_observations_from_project(proj)
-    obs = @list.observations.select { |o| check_permission(o) }
-    unless @user.projects_member.include?(proj)
-      obs.select! { |o| o.user == @user }
-    end
-    obs &= proj.observations
-    return unless obs.any?
-
-    proj.remove_observations(obs)
-    flash_notice(:removed_from_project.
-                    t(object: "#{obs.length} #{:observations.l}",
-                      project: proj.title))
-    @any_changes = true
-  end
-
-  def attach_images_to_project(proj)
-    imgs = @list.observations.map(&:images).flatten.uniq.
-           select { |i| check_permission(i) }
-    imgs -= proj.images
-    return unless imgs.any?
-
-    proj.add_images(imgs)
-    flash_notice(:attached_to_project.
-                    t(object: "#{imgs.length} #{:images.l}",
-                      project: proj.title))
-    @any_changes = true
-  end
-
-  def remove_images_from_project(proj)
-    imgs = @list.observations.map(&:images).flatten.uniq.
-           select { |i| check_permission(i) }
-    unless @user.projects_member.include?(proj)
-      imgs.select! { |i| i.user == @user }
-    end
-    imgs &= proj.observations
-    return unless imgs.any?
-
-    proj.remove_images(imgs)
-    flash_notice(:removed_from_project.
-                    t(object: "#{imgs.length} #{:images.l}",
-                      project: proj.title))
-    @any_changes = true
   end
 
   ##############################################################################
