+++ app/views/species_lists/observations/add_remove_observations.html.erb
@@ -1,15 +1,16 @@
 <%
-  @title = :species_list_add_remove_title.t
+# Linked from observations :index and species_lists :show
+@title = :species_list_add_remove_title.t
 
-  tabs = [
-    link_with_query(:species_list_add_remove_cancel.t,
-                    controller: :observations,
-                    action: :index)
-  ]
-  @tabsets = { right: draw_tab_set(tabs) }
+tabs = [
+  link_with_query(:species_list_add_remove_cancel.t,
+                  controller: :observations,
+                  action: :index)
+]
+@tabsets = { right: draw_tab_set(tabs) }
 %>
 
-<%= form_tag(add_query_param(action: :post_add_remove_observations)) do %>
+<%= form_with(url: add_query_param(action: :post_add_remove_observations)) do %>
   <p><%= :species_list_add_remove_body.tp(num: @query.num_results) %></p>
 
   <div class="form-group">
