+++ app/views/species_lists/observations/manage_species_lists.html.erb
@@ -35,8 +35,10 @@
 
         <td>
           <%= if check_permission(species_list)
-            link_to(:REMOVE.t, { action: :remove_observation_from_species_list,
-                    observation: @observation.id, species_list: species_list.id},
+            link_to(:REMOVE.t,
+                    { action: :remove_observation_from_species_list,
+                      observation: @observation.id,
+                      species_list: species_list.id},
                     data: { confirm: :are_you_sure.l })
           end %>
         </td>
@@ -64,8 +66,10 @@
 
         <td>
           <%= if check_permission(species_list)
-            link_to(:ADD.t, {action: :add_observation_to_species_list,
-                    observation: @observation.id, species_list: species_list.id})
+            link_to(:ADD.t,
+                    { action: :add_observation_to_species_list,
+                      observation: @observation.id,
+                      species_list: species_list.id})
           end %>
         </td>
